# contracts/openapi/esquilospeak-learning-v1.yaml
openapi: 3.1.1
info:
  title: EsquiloSpeak Learning API
  version: 0.5.0
  description: >-
    Mobile contract for identity/profile onboarding and the first learning
    vertical slice: resolve a guest or account learner, manage profile and
    consent, merge guest progress, request privacy workflows, discover a
    course, load a lesson, submit an append-only attempt, read explainable
    mastery and review schedules, and reconcile offline mutations. Account
    authentication remains owned by an external OIDC provider.
  contact:
    name: EsquiloSpeak
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/mobile/v1/media/{mediaId}:
    get:
      tags: [Advanced Learning]
      operationId: getLearningMedia
      summary: Stream an authenticated learning-media asset
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - name: mediaId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Identifier'
      responses:
        '200':
          description: WAV media bytes for the requested learning asset
          headers:
            ETag:
              schema:
                type: string
          content:
            audio/wav:
              schema:
                type: string
                format: binary
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/mobile/v1/advanced/pronunciation:
    post:
      tags: [Advanced Learning]
      operationId: assessPronunciation
      summary: Assess a short pronunciation sample without retaining raw voice
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clientRequestId, audioBase64, expectedText, locale]
              properties:
                clientRequestId:
                  type: string
                  format: uuid
                audioBase64:
                  type: string
                  contentEncoding: base64
                  maxLength: 2700000
                expectedText:
                  type: string
                  maxLength: 500
                locale:
                  $ref: '#/components/schemas/Locale'
              additionalProperties: false
      responses:
        '200':
          description: Derived pronunciation feedback
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvancedFeedbackResult'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/mobile/v1/advanced/writing:
    post:
      tags: [Advanced Learning]
      operationId: getWritingFeedback
      summary: Get safety-filtered writing feedback
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        $ref: '#/components/requestBodies/TextFeedbackRequest'
      responses:
        '200':
          description: Writing feedback
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvancedFeedbackResult'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/mobile/v1/advanced/conversation:
    post:
      tags: [Advanced Learning]
      operationId: getConversationFeedback
      summary: Continue a safety-filtered guided conversation
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        $ref: '#/components/requestBodies/TextFeedbackRequest'
      responses:
        '200':
          description: Conversation feedback
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvancedFeedbackResult'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /api/mobile/v1/engagement:
    get:
      tags: [Engagement]
      operationId: getEngagementStatus
      summary: Get streak, XP, achievements, and reminder preference
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current engagement state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EngagementStatus'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/mobile/v1/engagement/activities:
    post:
      tags: [Engagement]
      operationId: recordEngagementActivity
      summary: Record one idempotent learning activity
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clientEventId, eventType, xpAwarded]
              properties:
                clientEventId:
                  type: string
                  format: uuid
                eventType:
                  type: string
                  maxLength: 40
                xpAwarded:
                  type: integer
                  minimum: 0
                  maximum: 1000
              additionalProperties: false
      responses:
        '200':
          description: Updated engagement state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EngagementStatus'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/mobile/v1/engagement/notification-preference:
    put:
      tags: [Engagement]
      operationId: updateNotificationPreference
      summary: Configure local reminder delivery preferences
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationPreference'
      responses:
        '200':
          description: Updated reminder preference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreference'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        External OIDC account token or local-development guest token. Protected
        mobile operations require the learning scope and learner role; the
        actor_type claim must be guest or account.
  parameters:
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      description: >-
        Caller-provided correlation identifier. Invalid or omitted values are
        replaced by the server and the canonical value is returned in the
        response header and error traceId.
      schema:
        type: string
        pattern: '^[A-Za-z0-9._-]{1,64}$'
        maxLength: 64
...
  requestBodies:
    TextFeedbackRequest:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [clientRequestId, contentRef, input, locale]
            properties:
              clientRequestId:
                type: string
                format: uuid
              contentRef:
                type: string
                maxLength: 160
              input:
                type: string
                maxLength: 4000
              locale:
                $ref: '#/components/schemas/Locale'
            additionalProperties: false
...
  responses:
    ErrorResponse:
      description: Request could not be completed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    Identifier:
      type: string
      pattern: '^[a-z][a-z0-9]*(?:-[a-z0-9]+)*$'
      minLength: 2
      maxLength: 100
    Locale:
      type: string
      pattern: '^[A-Za-z]{2,3}(?:-[A-Za-z0-9]{2,8})*$'
      examples: [vi, en, pt-BR]
...
    AdvancedFeedbackResult:
      type: object
      required: [id, kind, feedback, provider, createdAt]
      properties:
        id:
          type: string
          format: uuid
        kind:
          type: string
          enum: [pronunciation, writing, conversation]
        transcript:
          type: string
        score:
          type: number
          minimum: 0
          maximum: 100
        feedback:
          type: object
          additionalProperties: true
        provider:
          type: string
        createdAt:
          type: string
          format: date-time
      additionalProperties: false
...
    NotificationPreference:
      type: object
      required: [enabled, locale]
      properties:
        enabled:
          type: boolean
        reminderTime:
          type: string
          format: time
        locale:
          $ref: '#/components/schemas/Locale'
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      additionalProperties: false
    EngagementStatus:
      type: object
      required:
        [currentStreak, longestStreak, xp, achievements,
         notificationPreference]
      properties:
        currentStreak:
          type: integer
          minimum: 0
        longestStreak:
          type: integer
          minimum: 0
        xp:
          type: integer
          minimum: 0
        lastLearningDate:
          type: string
          format: date
        achievements:
          type: array
          items:
            type: object
            required: [code, earnedAt]
            properties:
              code:
                $ref: '#/components/schemas/Identifier'
              earnedAt:
                type: string
                format: date-time
            additionalProperties: false
        notificationPreference:
          $ref: '#/components/schemas/NotificationPreference'
      additionalProperties: false
...
    ApiError:
      type: object
      required: [type, title, status, code, traceId, retryable]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          minLength: 1
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]+$'
        traceId:
          type: string
          minLength: 1
        retryable:
          type: boolean
        violations:
          type: array
          items:
            type: object
            required: [field, code]
            properties:
              field:
                type: string
              code:
                type: string
            additionalProperties: false
      additionalProperties: false